# example-microservice/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Delivery Specification
  description: This is a microservice for managing delivery in the YumYumNow company.
  version: 1.0.0
servers:
  - url: http://localhost:8080
...
paths:
  /order:
    get:
      summary: Retrieve a list of all orders
      operationId: getOrders
      tags:
        - Order
      parameters:
        - name: authorization
          in: query
          description: The userId to check if they have the rights to make this request
          required: true
          schema:
            type: integer
            format: int64
      description: return a list of order objects
      responses:
        '200':
          description: Successful response, orders received
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        '400':
          description: Unsuccessful, orders cannot be retrieved because of bad request
        '403':
          description: Unsuccessful, entity does not have access rights to retrieve all orders
        '404':
          description: Unsuccessful, no orders were found
  /order/{orderId}/courier/{courierId}:
    put:
      summary: Set courier id of this order
      operationId: setCourierId
      tags:
        - Order
      description: update the courier id of the order
      parameters:
        - name: orderId
          in: path
          description: id of the order to update
          required: true
          schema:
            type: integer
            format: int64
        - name: courierId
          in: path
          description: id of the courier to set
          required: true
          schema:
            type: integer
            format: int64
        - name: authorization
          in: query
          description: The userId to check if they have the rights to make this request
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Order object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
      responses:
        '200':
          description: Successful response, courier id of order set
        '400':
          description: Unsuccessful, courier id cannot be updated because of bad request
        '403':
          description: Unsuccessful, entity does not have access rights to update courier id
        '404':
          description: Unsuccessful, no order or courier id was found
  /order/{orderId}/preparation-time:
    put:
      summary: Set the preparation time of an order
      operationId: setDeliverTime
      tags:
        - Order
      description: update the preparation time of the order
      parameters:
        - name: orderId
          in: path
          description: id of the order to update
          required: true
          schema:
            type: integer
            format: int64
        - name: authorization
          in: query
          description: The userId to check if they have the rights to make this request
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Order object
        required: true
        content:
          application/json:
            schema:
              type: string
              format: time
      responses:
        '200':
          description: Successful response, preparation time of order set
        '400':
          description: Unsuccessful, preparation time cannot be updated because of bad request
        '403':
          description: Unsuccessful, entity does not have access rights to update preparation time
        '404':
          description: Unsuccessful, no preparation time was found
...
  /order/{orderId}:
    get:
      summary: Retrieve an order given the order id
      operationId: getOrder
      tags:
        - Order
      description: return the order corresponding to the id
      parameters:
        - name: orderId
          in: path
          description: id of the order to retrieve
          required: true
          schema:
            type: integer
            format: int64
        - name: authorization
          in: query
          description: The userId to check if they have the rights to make this request
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful response, order received
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Unsuccessful, order cannot be retrieved because of bad request
        '403':
          description: Unsuccessful, entity does not have access rights to retrieve order
        '404':
          description: Unsuccessful, no order was found
    post:
      summary: add an order given the order id
      operationId: makeOrder
      tags:
        - Order
      description: add the order corresponding to the id
      parameters:
        - name: orderId
          in: path
          description: id of the order to add
          required: true
          schema:
            type: integer
            format: int64
        - name: authorization
          in: query
          description: The userId to check if they have the rights to make this request
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Order object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
      responses:
        '200':
          description: Successful response, order added
        '400':
          description: Unsuccessful, order cannot be added because of bad request
        '403':
          description: Unsuccessful, entity does not have access rights to add order
        '404':
          description: Unsuccessful, no order was found
    put:
      summary: Update an order given the order id
      operationId: updateOrder
      tags:
        - Order
      description: update the order corresponding to the id
      parameters:
        - name: orderId
          in: path
          description: id of the order to update
          required: true
          schema:
            type: integer
            format: int64
        - name: authorization
          in: query
          description: The userId to check if they have the rights to make this request
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Order object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
      responses:
        '200':
          description: Successful response, order updated
        '400':
          description: Unsuccessful, order cannot be updated because of bad request
        '403':
          description: Unsuccessful, entity does not have access rights to update order
        '404':
          description: Unsuccessful, no order was found
...
  /order/{vendorId}:
    get:
      summary: Retrieve the next order that belongs to a given vendor
      operationId: getNextOrderForVendor
      tags:
        - Order
      description: return the next order object that is assigned to specific vendor
      parameters:
        - name: vendorId
          in: path
          description: id of the vendor to retrieve the order from
          required: true
          schema:
            type: integer
            format: int64
        - name: authorization
          in: query
          description: The userId to check if they have the rights to make this request
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful response, order received
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Unsuccessful, order cannot be retrieved because of bad request
        '403':
          description: Unsuccessful, entity does not have access rights to retrieve vendor order
        '404':
          description: Unsuccessful, no order was found
  /order/{orderId}/rating:
    get:
      operationId: getOrderRating
      tags:
        - Order
      summary: Retrieve the rating of a specific order
      description: Retrieve the rating details for a specific order
      parameters:
        - in: path
          name: orderId
          schema:
            type: integer
            format: int64
          required: true
          description: Id of the order to retrieve rating
        - name: authorization
          in: query
          description: The userId to check if they have the rights to make this request
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful response, rating number (out of 5) for the specific order received
          content:
            application/json:
              schema:
                type: number
                format: Long
        '404':
          description: Unsuccessful, no rating details found for the specified order
        '403':
          description: Unsuccessful, entity does not have access rights to retrieve order rating
    put:
      operationId: putOrderRating
      tags:
        - Order
      summary: Change the rating number (out of 5)
      description: Update both of the orders ratings
      parameters:
        - in: path
          name: orderId
          schema:
            type: integer
            format: int64
          required: true
          description: Id of the order to update rating
        - name: authorization
          in: query
          description: The userId to check if they have the rights to make this request
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: The rating number to update the rating field of an order
        required: true
        content:
          application/json:
            schema:
              type: number
              format: Long
      responses:
        '200':
          description: Successful update of rating
        '400':
          description: Invalid arguments offered
        '404':
          description: OrderId was not found
        '401':
          description: Unauthorized
        '403':
          description: Unsuccessful, entity does not have access rights to update order rating
  /order/{orderId}/pickup-destination:
    get:
      operationId: getPickupDestination
      summary: Retrieve vendor location of the order
      tags:
        - Order
      description: Return the vendor location of the specified order
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: authorization
          in: query
          description: The userId to check if they have the rights to make this request
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful response, vendor location of the order received
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'
        '400':
          description: Unsuccessful, vendor location of the order cannot be retrieved because of a bad request
        '403':
          description: Unsuccessful, entity does not have access rights to retrieve vendor location
        '404':
          description: Unsuccessful, vendor location for the order not found
  /order/{orderId}/final-destination:
    get:
      operationId: getFinalDestination
      summary: Get the final destination of a specific order
      tags:
        - Order
      description: Retrieve the final destination of an order
      parameters:
        - name: authorization
          in: query
          description: The userId to check if they have the rights to make this request
          required: true
          schema:
            type: integer
            format: int64
        - in: path
          name: orderId
          schema:
            type: integer
            format: int64
          required: true
          description: Id of the order to get its final destination
      responses:
        '200':
          description: Successful response, order found and final destination can be retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'
        '403':
          description: Unsuccessful, entity does not have access rights to retrieve final destination
        '404':
          description: Unsuccessful, order not found by id
...
components:
  schemas:
    Order:
      type: object
      x-class-extra-annotation: "@javax.persistence.Entity @javax.persistence.Table(name = \"order_table\") @lombok.AllArgsConstructor @lombok.NoArgsConstructor"
      x-field-extra-annotation: "@javax.persistence.OneToOne(cascade = javax.persistence.CascadeType.ALL) @javax.persistence.JoinColumn(name =\"orderId\")"
      properties:
        status:
          type: string
          enum:
            - pending
            - accepted
            - rejected
            - preparing
            - given-to-courier
            - in-transit
            - delivered
          description: the current status of the order expressed as one of the above options
        id:
          type: integer
          format: int64
          description: the unique id of an order
          x-field-extra-annotation: "@javax.persistence.Id"
        courierId:
          type: integer
          format: int64
          example: 88
          description: the id of the courier that is delivering the order
        vendorId:
          type: integer
          format: int64
          example: 99
          description: the id of the vendor that prepares the order
        deliveryDestination:
          $ref: '#/components/schemas/Location'
        timeValues:
          $ref: '#/components/schemas/Time'
        ratingNumber:
          type: number
          format: Long
          description: a number out of five (0-5 stars)
...
    Location:
      description: Any kind of address expressed as a longitude and latitude
      type: object
      x-class-extra-annotation: "@javax.persistence.Embeddable"
      x-field-extra-annotation: "@javax.persistence.Embedded"
      properties:
        latitude:
          type: number
          format: float
          example: 23.92488
        longitude:
          type: number
          format: float
          example: 74.7099
    Time:
      description: Collection of time attributes of an order
      type: object
      x-class-extra-annotation: "@javax.persistence.Embeddable"
      x-field-extra-annotation: "@javax.persistence.Embedded"
      properties:
        prepTime:
          type: string
          format: time
          description: the time it takes to prepare the order in format "HH:MM:SS"
          example: '01:30:00'
        expectedDeliveryTime:
          type: string
          format: date-time
          description: the time the order will be planned to be delivered
        actualDeliveryTime:
          type: string
          format: date-time
          description: the time the order actually is delivered
        orderTime:
          type: string
          format: date-time
          description: the time the order was placed
...